JP-1168 Update NIRCam level 2a regression test
